Use _fh() instead of fh() as fh() does not take arguments
The first call to get_Seq_by_id always works. But a second call to the method do not return the correct sequence. The method fh() does not take any argument. It is the method _fh() that needs to be used in order to set the filehandle to a new value. When the filehandle is updated using _fh(), the correct sequence is return for any other call of get_Seq_by_id.